Fix of << for input ranges, addition of even, pow and square functors
Operator` <<` for ranges was broken for *InputRange*s, because they are not `sized_range`s. Fixed by removing a check for `std::ranges::empty`.

Added new functors for mathematical operations - a predicate `even` (self-explanatory) and `square`, which is a "specialisation" of `pow`, which raises an argument to a given power. Those names are not set in stone - further analysis for potentially better names is required.